requirements.txt

pytz
Pillow
requests
